fix: split results.py into storage and collection modules
Split results.py (1730 LOC) into modular components for better
code organization:

- storage.py (660 LOC): DistributionFitResult, LazyMetricsContext,
  constants, and type aliases
- collection.py (1194 LOC): BaseFitResults, EagerFitResults,
  LazyFitResults, and factory functions
- results.py (99 LOC): Re-exports for backward compatibility

Added type aliases:
- ContinuousHistogram: Tuple for (density_values, bin_edges)
- DiscreteHistogram: Tuple for (x_values, empirical_pmf)

Extracted named constants:
- FITTING_SAMPLE_SIZE, DEFAULT_PVALUE_THRESHOLD, DEFAULT_KS_THRESHOLD
- DEFAULT_AD_THRESHOLD, DEFAULT_BINS, DEFAULT_BOOTSTRAP_SAMPLES
- DEFAULT_MAX_SAMPLES, DEFAULT_DPI, DEFAULT_SAMPLE_SIZE

All existing imports from spark_bestfit.results continue to work.
